This commit introduces a powerful new framework for time series detection using MATLAB's Deep Learning Toolbox.
Two new classes have been added:
1.  `vlt.signal.timeseriesDetectorML.dlt`: A general-purpose wrapper class that can train and evaluate any deep learning network defined using the toolbox. It features an efficient, batched `evaluateTimeSeries` method that processes the time series in chunks for high performance.
2.  `vlt.signal.timeseriesDetectorML.conv1dNet`: A user-friendly convenience class that inherits from `dlt`. It simplifies the creation of a common 2-tier 1D CNN by accepting high-level architectural parameters in its constructor, abstracting away the manual layer definition.

This commit also includes:
- `dltDemo.m` and `conv1dNetDemo.m` to showcase how to use the new classes.
- `testDlt.m` and `testConv1dNet.m` with unit tests to verify the functionality of the new classes.
- A restored `base.m` file containing essential helper functions.
- Fixes to the `dlt.evaluateTimeSeries` method to resolve a `predict` function error by using a robust `for` loop to build the batch matrix, avoiding data type issues with `im2col`.
- A fix to the `DetectorSamples` property in the `dlt` and `base` classes, changing the type to `double` to prevent `padarray` errors.

+++ +vlt/+signal/+timeseriesDetectorML/base.m
@@ -17,10 +17,10 @@ classdef (Abstract) base
                 timeSeriesTimeStamps (:,1) double
                 timeSeriesData (:,1) double
                 detectedTimeStamps (1,:) double
-                detectorSamples (1,1) uint64
+                detectorSamples (1,1) double
                 examplesArePositives (1,1) logical = true
                 options.optimizeForPeak (1,1) logical = false
-                options.peakFindingSamples (1,1) uint64 = 10
+                options.peakFindingSamples (1,1) double = 10
                 options.useNegativeForPeak (1,1) logical = false
             end
